requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
djoser~=2.3.1
python-dotenv~=1.0.1